requirements.txt
scikit-learn
xgboost
streamlit
matplotlib 
seaborn 
openai 
python-dotenv 
streamlit 
pandas 
numpy
